+++ examples/03-dev/dist-net-all.js
@@ -3,8 +3,8 @@ require('../../util/change_cwd_to')(__dirname + '/demo');
 
 var projectConfig = require('./demo/lila.config');
 projectConfig.htmlAbsoluteAndCdnPath = !0;
-projectConfig.networkOptions = [];
-projectConfig.networkOptions[100] = {
+projectConfig.network = [];
+projectConfig.network[100] = {
     staticDomain: 'http://www.baidu.com/lila',
     cssAbsolutePathPrefix: '/lila',
     backupHtml: !0