+++ examples/03-dev/dist-net-cus-all.js
@@ -5,8 +5,8 @@ require('./cus');
 
 var projectConfig = require('./demo/lila.config');
 projectConfig.htmlAbsoluteAndCdnPath = !0;
-projectConfig.networkOptions = [];
-projectConfig.networkOptions[100] = {
+projectConfig.network = [];
+projectConfig.network[100] = {
     staticDomain: 'http://www.baidu.com/lila',
     cssAbsolutePathPrefix: '/lila',
     backupHtml: !0